moved dynamic bootstrap's test case to base test class
